Update RecommendationSystem.py
There is still a mistake with using recommend function after this one. If Mary always rates 1 more that John they will score a perfect correlation(1.0). But the recommend function does not account for the difference in their ratings. To do this the average of each user to the main user must be taken into account.